test/x/fleet-builder/requirements.txt
azure-storage-queue
azure-identity
